Reorganized dir struct
