+++ frontend/src/App.tsx
@@ -1,9 +1,9 @@
 import '@bcgov/design-tokens/css/variables.css';
 import '@bcgov/bc-sans/css/BC_Sans.css';
 import StyledApp from './App.style';
-import LandingPage from './Views/LandingPage';
-import Header from './Commons/Header';
-import Footer from './Commons/Footer';
+import LandingPage from './Views/LandingPage/LandingPage.tsx';
+import Header from './Commons/Header/Header.tsx';
+import Footer from './Commons/Footer/Footer.tsx';
 
 const App = () => (
   <StyledApp>
