+++ frontend/src/Commons/Footer/Footer.tsx
@@ -2,7 +2,7 @@
  * @desc A footer with links to our terms of agreement and BC NMP Help resource
  *@author @GDamaso
  */
-import StyledFooter from './Styles/Footer.style';
+import StyledFooter from './Footer.style';
 
 const termsURL = 'https://nmp.apps.nrs.gov.bc.ca/Information/Disclaimer';
 const helpURL = 'https://www2.gov.bc.ca/gov/content/industry/agriculture-seafood/agricultural-land-and-environment/soil-nutrients/nutrient-management/nutrient-management-calculator/support-for-the-bc-nutrient-management-calculator';
