+++ frontend/src/Views/MainPage/MainPage.styles.ts
